feat(supabase): add swizzle support to supabase
