+++ .changeset/perfect-donuts-grab.md
@@ -1,0 +1,9 @@
+---
+"@refinedev/supabase": minor
+---
+
+feat: added refine.config.js to support swizzling. Now with swizzle support, you can easily customize supabase data provider for your needs.
+
+feat: tests added for utility functions.
+
+chore: utility functions have been moved to their own files.
