+++ packages/supabase/src/utils/handleError.ts
@@ -1,0 +1,11 @@
+import { HttpError } from "@refinedev/core";
+import { PostgrestError } from "@supabase/supabase-js";
+
+export const handleError = (error: PostgrestError) => {
+    const customError: HttpError = {
+        ...error,
+        message: error.message,
+        statusCode: parseInt(error.code),
+    };
+    return Promise.reject(customError);
+};
